Renamed the plugins dir as toyide.
